*Move SortTaskByDateCommand from persons folder into tasks folder
* Update DG with SortTaskByDateCommand and diagram
