+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -9,7 +9,7 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.logic.parser.Coach2K22Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
