+++ src/main/java/seedu/address/logic/commands/person/AddPersonTagCommand.java
@@ -12,7 +12,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.name.Name;
 import seedu.address.model.note.Note;
