+++ src/main/java/seedu/address/logic/commands/person/EditPersonCommand.java
@@ -18,7 +18,7 @@ import seedu.address.logic.EditPersonDescriptor;
 import seedu.address.logic.EditTaskDescriptor;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.name.Name;
 import seedu.address.model.note.Note;
