+++ src/main/java/seedu/address/logic/commands/strategy/LoadCourtCommand.java
@@ -7,7 +7,7 @@ import java.io.File;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.image.Image;
