+++ src/main/java/seedu/address/logic/commands/task/AddTaskCommand.java
@@ -13,7 +13,7 @@ import java.util.Set;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.name.Name;
 import seedu.address.model.person.Person;
