+++ src/main/java/seedu/address/logic/commands/task/ClearTaskCommand.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.TaskBook;
 import seedu.address.model.task.Date;
