+++ src/main/java/seedu/address/logic/commands/task/SortTaskByDateCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.person;
+package seedu.address.logic.commands.task;
 
 import static java.util.Objects.requireNonNull;
 
@@ -8,7 +8,7 @@ import java.util.List;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.TaskBook;
 import seedu.address.model.task.Task;
