+++ src/main/java/seedu/address/logic/commands/task/exceptions/CommandException.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.exceptions;
+package seedu.address.logic.commands.task.exceptions;
 
 /**
  * Represents an error which occurs during execution of a {@link Command}.
