+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -19,7 +19,7 @@ import java.util.List;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.EditPersonDescriptor;
 import seedu.address.logic.EditTaskDescriptor;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.name.PersonNameContainsKeywordsPredicate;
