+++ src/test/java/seedu/address/logic/commands/task/EditTaskCommandTest.java
@@ -21,7 +21,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.EditTaskDescriptor;
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.task.exceptions.CommandException;
 import seedu.address.logic.commands.person.ClearPersonCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
