+++ src/test/java/seedu/address/logic/commands/task/SortTaskByDateCommandTest.java
@@ -14,7 +14,6 @@ import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.person.SortTaskByDateCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
